refactor packet.command
