+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/transport/mysql/packet/command/api/impl/QueryResponsePackets.java
@@ -15,9 +15,10 @@
  * </p>
  */
 
-package io.shardingsphere.proxy.transport.mysql.packet.command.reponse;
+package io.shardingsphere.proxy.transport.mysql.packet.command.api.impl;
 
 import io.shardingsphere.proxy.transport.mysql.constant.ColumnType;
+import io.shardingsphere.proxy.transport.mysql.packet.command.api.CommandResponsePackets;
 import io.shardingsphere.proxy.transport.mysql.packet.command.text.query.ColumnDefinition41Packet;
 import io.shardingsphere.proxy.transport.mysql.packet.command.text.query.FieldCountPacket;
 import io.shardingsphere.proxy.transport.mysql.packet.generic.EofPacket;
