+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/transport/mysql/packet/command/text/initdb/ComInitDbPacket.java
@@ -20,9 +20,9 @@ package io.shardingsphere.proxy.transport.mysql.packet.command.text.initdb;
 import io.shardingsphere.core.constant.ShardingConstant;
 import io.shardingsphere.proxy.transport.mysql.constant.ServerErrorCode;
 import io.shardingsphere.proxy.transport.mysql.packet.MySQLPacketPayload;
-import io.shardingsphere.proxy.transport.mysql.packet.command.CommandPacket;
-import io.shardingsphere.proxy.transport.mysql.packet.command.CommandPacketType;
-import io.shardingsphere.proxy.transport.mysql.packet.command.reponse.CommandResponsePackets;
+import io.shardingsphere.proxy.transport.mysql.packet.command.api.CommandPacket;
+import io.shardingsphere.proxy.transport.mysql.packet.command.api.CommandPacketType;
+import io.shardingsphere.proxy.transport.mysql.packet.command.api.CommandResponsePackets;
 import io.shardingsphere.proxy.transport.mysql.packet.generic.ErrPacket;
 import io.shardingsphere.proxy.transport.mysql.packet.generic.OKPacket;
 import lombok.Getter;
